requirements.txt
click>=8.0.0
rich>=13.0.0
pytest>=7.0.0
pytest-cov>=4.0.0
textual>=0.50.0
pyyaml>=6.0
